src/requirements.txt
Flask==3.0.3
requests-oauthlib==2.0.0
google-auth==2.35.0
google-cloud-bigquery==3.25.0
google-cloud-storage==2.17.0
google-cloud-aiplatform==1.59.0
pandas==2.2.1
streamlit==1.36.0
streamlit-option-menu==0.3.13
plotly==5.23.0
seaborn==0.13.2
altair==4.2.2
statsmodels==0.14.2
matplotlib==3.9.0
scikit-learn==1.5.0
numpy==1.26.4
protobuf==4.25.3
db-dtypes==1.3.0
kfp==2.7.0
google-cloud-pipeline-components==2.16.1
python-dotenv==1.0.1
Flask-Session==0.8.0
mysqlclient==2.2.4
flask_sqlalchemy=3.1.1 
greenlet=3.1.1 
sqlalchemy=2.0.35
